backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
xai-sdk
python-multipart
pillow
python-dotenv